+++ source/Examples/Resources/CMBoundaryGroups/CMBoundaryGroups_Present.ps1
@@ -1,6 +1,7 @@
 <#
     .SYNOPSIS
-        A DSC configuration script to add a boundary group and match boundaries in Configuration Manager.
+        A DSC configuration script to add a boundary group and match boundaries in
+        the boundary group in Configuration Manager.
 #>
 Configuration Example
 {
@@ -10,10 +11,26 @@ Configuration Example
     {
         CMBoundaryGroups ExampleSettings
         {
-            SiteCode      = 'Lab'
-            BoundaryGroup = 'TestGroup'
-            Boundaries    = 'TB1','TB2'
-            Ensure        = 'Present'
+            SiteCode       = 'Lab'
+            BoundaryGroup  = 'TestGroup'
+            Boundaries     = @(
+                DSC_CMBoundaryGroupsBoundaries
+                {
+                    Value = '10.1.1.1/24'
+                    Type  = 'IPSubnet'
+                }
+                DSC_CMBoundaryGroupsBoundaries
+                {
+                    Value = '10.1.1.1-10.1.1.255'
+                    Type  = 'IPRange'
+                }
+                DSC_CMBoundaryGroupsBoundaries
+                {
+                    Value = 'First-Site'
+                    Type  = 'AdSite'
+                }
+            )
+            BoundaryAction = 'Match'
         }
     }
 }